Fix packer package installation issues
Not entirely sure what happened here. I think it may have just been
some upstream issues. Adding the repositories shouldn't have made a
change, and the change didn't actually fix anything until the 2nd or
third time I ran it. So, this may make a difference and may not. But,
it's working now, so that's good